+++ src/main/java/racingcar/RacingCarController.java
@@ -1,21 +1,14 @@
 package racingcar;
 
-import java.util.InputMismatchException;
-import java.util.Scanner;
-
 public class RacingCarController {
 
-    private final Scanner scanner;
+    private final InputView inputView;
 
-    public RacingCarController(Scanner scanner) {
-        this.scanner = scanner;
+    public RacingCarController(InputView inputView) {
+        this.inputView = inputView;
     }
 
     public void run() {
-        try {
-            int tryCount = scanner.nextInt();
-        } catch (InputMismatchException e) {
-            throw new NumberFormatException("시도횟수는 숫자이어야 합니다.");
-        }
+        int tryCount = inputView.inputTryCount();
     }
 }
